my_blog_project/requirements.txt


# These are the core packages required for your Django project
django==3.2
django-cms  
mysqlclient  #Necessray for web sercie to make rquest towards db sercie (see docker-compose.yaml file)

# Additional packages for Django CMS and related functionality [ Note tha you can override most plugin:see plugin github]
django-staticfiles              # For handling static files
djangocms-style                 # For styling in Django CMS
djangocms-text-ckeditor         # For text editing with CKEditor in Django CMS
djangocms-file                  # add files and entiere folder to you page
django-filer                    # handling of files and images
djangocms-admin-style           # to display admin page with beautifull style 
djangocms-link                  # Add link to my page 
django-select2                  # add internal link plugin 
djangocms-picture               # to add image
djangocms-style                 # create HTML containers with classes, styles, ids and other attributes
djangocms-snippet               # plugin to place custom markup or Javascript on pages
djangocms-audio                 # plugin for audio ['mp3', 'ogg', 'wav']
djangocms-video                 # plugin for video ['mp4', 'webm', 'ogv']
djangocms-googlemap             # add googlemap plugin
easy-thumbnails
# djangocms-forms <-- souleve des erreurs  ImportError: cannot import name 'ugettext_lazy' from 'django.utils.translation' (/usr/local/lib/python3.8/site-packages/django/utils/translation/__init__.py),
djangocms-column                # for multi column plugin
# djangocms-forms
django-ckeditor
# django-ckeditor-uploader

# django-parler                  

# djangocms-categorization                    #  To add Category
djangocms-aldryn-categories==2.0.1          #  To add Category

#  LIVE-RELOAD : reload after modi of *.html, *.py, *.sh
watchdog
